+++ src/starkware/algebra/lde/cached_lde_manager.cc
@@ -63,7 +63,7 @@ void CachedLdeManager::EvalAtPoints(
     gsl::span<const FieldElementSpan> outputs) {
   ASSERT_RELEASE(done_adding_, "Must call FinalizeAdding() before calling EvalAtPoints()");
   ASSERT_RELEASE(outputs.size() == n_columns_, "Wrong number of output columns");
-  for (size_t column_index = 0; column_index < n_columns_; ++column_index) {
+  for (uint64_t column_index = 0; column_index < n_columns_; ++column_index) {
     ASSERT_RELEASE(
         coset_and_point_indices.size() == outputs[column_index].Size(),
         "Number of output points is different than number of input points");
@@ -75,12 +75,12 @@ void CachedLdeManager::EvalAtPoints(
 
   if (config_.store_full_lde) {
     // Look up coset in cache.
-    for (size_t i = 0; i < coset_and_point_indices.size(); ++i) {
+    for (uint64_t i = 0; i < coset_and_point_indices.size(); ++i) {
       const auto& [coset_index, point_index] = coset_and_point_indices[i];
       ASSERT_RELEASE(
           cache_[coset_index].has_value(),
           "EvalAtPoints with config_.store_full_lde requested a coset that is not cached!");
-      for (size_t column_index = 0; column_index < n_columns_; ++column_index) {
+      for (uint64_t column_index = 0; column_index < n_columns_; ++column_index) {
         outputs[column_index].Set(i, (*cache_[coset_index])[column_index].At(point_index));
       }
     }
@@ -89,8 +89,8 @@ void CachedLdeManager::EvalAtPoints(
 
   // If not found, consider computing the entire coset.
   if (config_.use_fft_for_eval) {
-    std::map<uint64_t, std::vector<size_t>> coset_to_query_index;
-    for (size_t i = 0; i < coset_and_point_indices.size(); ++i) {
+    std::map<uint64_t, std::vector<uint64_t>> coset_to_query_index;
+    for (uint64_t i = 0; i < coset_and_point_indices.size(); ++i) {
       const auto& [coset_index, point_index] = coset_and_point_indices[i];
       (void)point_index;  // Unused.
       coset_to_query_index[coset_index].push_back(i);
@@ -98,8 +98,8 @@ void CachedLdeManager::EvalAtPoints(
     LdeCacheEntry entry = InitializeEntry();
     for (const auto& [coset_index, query_indices] : coset_to_query_index) {
       auto coset_evaluation = EvalOnCoset(coset_index, &entry);
-      for (size_t query_index : query_indices) {
-        for (size_t column_index = 0; column_index < n_columns_; ++column_index) {
+      for (uint64_t query_index : query_indices) {
+        for (uint64_t column_index = 0; column_index < n_columns_; ++column_index) {
           const uint64_t point_index = coset_and_point_indices[query_index].second;
           outputs[column_index].Set(query_index, (*coset_evaluation)[column_index].At(point_index));
         }
@@ -118,13 +118,13 @@ void CachedLdeManager::EvalAtPoints(
     points.PushBack(domain->GetFieldElementAt(point_index));
   }
 
-  for (size_t column_index = 0; column_index < n_columns_; ++column_index) {
+  for (uint64_t column_index = 0; column_index < n_columns_; ++column_index) {
     EvalAtPointsNotCached(column_index, points, outputs[column_index]);
   }
 }
 
 void CachedLdeManager::EvalAtPointsNotCached(
-    size_t column_index, const ConstFieldElementSpan& points, const FieldElementSpan& output) {
+    uint64_t column_index, const ConstFieldElementSpan& points, const FieldElementSpan& output) {
   ASSERT_RELEASE(
       lde_manager_.HasValue(), "Cannot evaluate new values after FinalizeEvaluations() was called");
   lde_manager_->EvalAtPoints(column_index, points, output);
@@ -134,7 +134,7 @@ CachedLdeManager::LdeCacheEntry CachedLdeManager::InitializeEntry() const {
   LdeCacheEntry entry;
   const uint64_t coset_size = domain_size_;
   entry.reserve(n_columns_);
-  for (size_t i = 0; i < n_columns_; ++i) {
+  for (uint64_t i = 0; i < n_columns_; ++i) {
     entry.push_back(
         FieldElementVector::MakeUninitialized(coset_offsets_->At(0).GetField(), coset_size));
   }
